Switch from Haversine formula to earth_distance Postgres extension
More readible and solves the issues we had where the query lat/lng
resulting in a value outside of [1,-1] within ACOS() leading to an
error, "input is out of range". This appears to occur when the lat/lng
in the query exactly matches a thing's location. E.g.

```
SELECT
  ACOS(
    COS(RADIANS(37.7876114900295))
    * COS(RADIANS(37.78761149002950))
    * COS(RADIANS(-122.39915856367400) - RADIANS(-122.399158563674))
    + SIN(RADIANS(37.7876114900295))
    * SIN(RADIANS(37.78761149002950))
  )
;
```

Results in `ERROR:  input is out of range`.

I _think_ this is because of floating point imprecision because if we
remove the ACOS:

```
SELECT
  COS(RADIANS(37.7876114900295))
  * COS(RADIANS(37.78761149002950))
  * COS(RADIANS(-122.39915856367400) - RADIANS(-122.399158563674))
  + SIN(RADIANS(37.7876114900295))
  * SIN(RADIANS(37.78761149002950))
;
```

We get `1` which should give an `ACOS(1)` == `0`.